Add -d/--output-dep-path options for "-M" deps.
Both -d and --output-dep-path are used to specify where the
dependency files should be generated (for a compilation that
is probably using -o to specify an output .bc directory).
Omitting -d/--output-dep-path will result in the output .d
files being placed in the current directory.

Sample corresponding commands for compilation/dep generation:
$ slang -o bc            foo.rs
    creates bc/foo.bc
$ slang -o bc -M -d deps foo.rs
    creates deps/foo.d that contains dependencies for bc/foo.bc

Change-Id: I05b87a0896f5bb3ac06db094a6c55e58fea57806