Update error diagnostic for struct/array copy.
An upstream Clang change caused F_struct_array_copy to start failing with the
latest downstreaming work. I noticed that the change makes the diagnostic more
clear, but can be confusing for the StartLoc (for the variable). This change
updates the test to use the new source location (for the actual assignment
operator), while making sure that StartLoc points to the LHS variable in the
assignment.

Change-Id: Ie58b602ad4ce69d80b82fc03475fbd5d4beb17c7