Simpler ScriptC constructors
This change replaces the old 3-argument version of reflected ScriptC
constructors with a new single argument (just the RenderScript object).
We can do this safely because there are lookup functions for Resources
and resource IDs directly in the Android API (and accessible via the
RenderScript object). For now, we are still reflecting the legacy version
of the constructor as well, since all existing app source code would be
using it.

Change-Id: I362d6e2bea0f580630b5da9049df029b0f8d7620