Add BitWriter_3_2 for JB+ target API.
This change switches llvm-rs-cc to use a single format for JB+ target API
LLVM bitcode. This simplifies upstream rebases, considering the divergent
compressed bitcode format that will be present in LLVM 3.3. We may move
back to using the upstream BitWriter again at some point in the future, but
for now, it seems like the best choice is to stick with the 3.2 format.

Change-Id: I8cfc54d7da7c7168f8f0d5f8fc24c2598b81aff8