Remove dead code trying to write opt-level
The previous commit exposed a piece of dead code, that apperently
tried to store the optimization level in the LLVM module generated
by the slang backend. As it was not functional and there also does
not seem to be a good reason to do so, we remove this code.

No functional change intended.

Change-Id: Ieffb1dc3e48db5aabe6ecab46bd12132161c694d