Fix llvm-rs-cc tests due to header changes.
We moved graphics types to the separate rs_graphics.rsh header, so we need
to include it when we want to use rs_font, etc. in our tests. Some tests
were modified to use more relevant RS object types, since we don't want
to rely on graphics types going forward.

Change-Id: I4a85d183747f980670b377c9df8b63c2bee0e164