Added in fix to allow generation of valid debug info
Changed the bitcode writer of the llvm-rs-cc compiler
to use the current version of llvm when generating debug
metadata. If the debug info is generated with the current
fixed version writers the resulting bitcode is invalid and
will fail verification in the bcc compiler. However, this
means that, currently, only revisions which have compatible
versions of llvm will work. The incompatibility is in the
debug info generation in slang and the verification/backend
in bcc. If bcc is expecting the debug info to be in a different
format that slang outputs the verification stage will fail. As
things stand only the newest revision will work, as the debug
info in AOSP LLVM was updated a few weeks ago. If the debug
bitcode is ahead of the reader on the device it will fail the
bitcode verification stage in bcc.

Change-Id: Iddb7adfe6639900f9917cb05941378019ecc37ba
Signed-off-by: Stephen McGroarty <stephen@codeplay.com>