Always set CodeGenOpts.StackRealignment to 1.
Bug: 22718149

Clang's driver usually passes -mstackrealign directly to cc1 (unless the
-mno-stackrealign flag was passed to it). If -mstackrealign isn't passed
to cc1, Clang will automatically generate "no-realign-stack" attributes
on all functions, which precludes modifying the stack pointer for any
aligned operations. This was an unexpected behavior that breaks
alignment for some vector loads/stores. We thus force StackRealignment
to 1 in order to prevent this unexpected and broken behavior.

external/clang/test/Driver/stackrealign.c shows the cc1 flags being
passed along for a given invocation of top-level clang. r145909 is the
commit that explains how this is mostly supposed to work.

Change-Id: I7f28ee5273e35387ff689bafe180a70ef3c13750