Restore missing min and max API.
min(float2/3/4, float) and max(float2/3/4, float) were defined in our
headers in K.  In L, they were removed by mistake.  This continued to
work however because a float would be auto converted by the compiler to
a float4.

In M, we started validating for unexpected external references, which
_Z3minDv4_ff is.  This can prevent programs compiled with K from running
on M.

This CL adds the missing APIs to the tests.

Bug:23389429

Change-Id: Idfece91a9e3ca6a53ac2e40e35f5d539c12b699a