Bump version to 2310 to account for RS object change.
Bug: 22926131

LLVM internally has an assertion that fails when trying to read the name
of an anonymous (literal) struct. We unfortunately have relied on
reading this name for the entire history of RenderScript, so that
prevents us from running an assertion-enabled build of LLVM currently.
Although we can't change the behavior on general builds (due to needing
the ability to read legacy bitcode), we want to prevent the
assertion-enabled build from crashing spuriously.

To enable that, https://android-review.googlesource.com/#/c/168294/
updates frameworks/rs such that our RS object types are no longer just
typedefs of anonymous structs. This allows the assertion build to run
correctly with modern compiled code, but won't allow it to work with
legacy bitcode (from compilers using previous headers). Since the
llvm-rs-cc version number is part of the bitcode wrapper, we can use it
to help bcc detect this invalid use of legacy code with our assertion
builds.

This change also updates a test that checked that the proper structure
type is being used (now that it can be more specific). The frameworks/rs
change listed above altered the headers to use the non-anonymous
version of the structs.

Change-Id: Ie16e8a9a71da3766973a2e43a8aec845adb1cc4d