Fix unnecessarily broken sections in slang tests.
These tests mostly were including broken kernel functions (in addition
to their other errors). It is best to remove these unnecessary blocks
for clarity in the individual tests. llvm-rs-cc is intended to return
C99-related errors first, and any possible RS-kernel-reflection errors
only once that completes successfully (which is why we never noticed
this before). I also renamed some of the misnamed tests (we don't retain
the P_ or F_ prefix on the source filenames).

Change-Id: I9995e8b05817c895676e0eaa1adde263dc639cc0