Fixed dummy root creation
There should be zero exported foreach functions in the metadata, if
there is no root function seen and there is no non-root kernel seen.

Create a dummy root only if there exists other non-root kernels.
Remove the nullptr placeholder in the exported for each function
vector if there is no kernel (root or not) seen in the script.

Also adjusted the lit test script to make it run on a setup where
the out directory is not under the source tree.

Change-Id: I4c4a5fe54a7fe926db9a432c9fca2ddc5f354ba8