Minor miscellaneous cleanup to type export facility.
(This is needed by general reduction.)

DETAILS:

Make certain functions const-clean.

TypeExportable() needs to recognize that it cannot export function
types, rather than asserting if it sees them.

Add comments to gReflectionTypes[] for readability.

When attempting RSExportType::Create(), allow a variable "containing"
the type (not "of" the type) to be specified.  This is important
because certain error message reporting absolutely requires the
variable to be known -- such as when
ConstantArrayTypeExportableHelper() wants to report something like
"multidimensional arrays cannot be exported".

Change-Id: Iff49effdf7cb71184af5846217fde5a493ce61e9