Remove array subrange (x1, x2) general reduction methods.
The current implementation is functionally incorrect.

A general reduction method that takes array inputs is implemented by
creating a temporary Allocation for each input and initializing the
Allocation from the input array.

There is no efficient way to implement the array subrange methods
today, because there is no efficient way to initialize an Allocation
from a subrange of an array (see http://b/26804454).  We would have to
do something like:
- Extend the RenderScript Java API; or
- For each input array, create a temporary array, populate the temporary
  array from the appropriate subrange of the the original array, and
  initialize a temporary Allocation from the temporary array instead of the
  input array; or
- For each input array, initialize an Allocation from the entire array,
  but use LaunchOptions to specify that only a subrange of the Allocation
  should be processed by the kernel.

Bug: 23535724
Change-Id: Ie9d3ebe445724512bb196aa661a295f217145918