Initialize temp var for a return value
Bug: 27698337

This is the temporary variable introduced for the expression used in
a return statement.

We use rsSetObject() to set the value of the temporary variable to
the original expression from the return statement, and to bump up
the sysRefCount of the temp.

However, to have rsSetObject() work properly, the temp has to be zero
initialized.

Change-Id: I0efb83fc7d54e95bb4d7a956f3f82b6d85188390
(cherry picked from commit 5767c359bac779b5d76d668ca5699fd5a0e3efc5)